moved ModelStatusAlert component into it's own file for better separation of concerns. #113.
